feat(sdk-spec): Krok 4a — strict canonical CBOR validator (Faza 6 deterministic profile, RFC 8949 §4.2.1 + f64-only) + MAX_NESTING_DEPTH=64 anti-DoS + usize::try_from for wasm32 + f32→f64 migration (GraphEdge.weight, FileMeta.upload_progress, AspectRatio::Custom.ratio, Viewport.density)

+++ tentaflow-sdk-spec/src/protocol/ui/schema/data_unions.rs
@@ -703,7 +703,7 @@ pub const ASPECTRATIO_UNION: UnionMeta = UnionMeta {
             rust_name: "Custom",
             wire_kind: "custom",
             fields: &[
-                FieldMeta { key: 0, name: "ratio", wire: "f32", required: true, default: None },
+                FieldMeta { key: 0, name: "ratio", wire: "f64", required: true, default: None },
             ],
         },
     ],
